feat(daemon): enhance Hermes WS stream collection and output clipping (#16)
- Extract clipSummary and buildWsUrl to daemon-utils.ts with testable contract
- Aggregate payload.text on turn.end/message.complete alongside incremental deltas
- Verified with monorepo build and lint
